requirements.txt
pyzmq>=25.0.0
numpy>=1.24.0
qiskit>=2.0.0
qiskit-aer>=0.14.0
